pyproject.toml
[project]
name = "xaoc2"
version = "0.1.0"
description = ""
authors = [{ name = "phillychi3", email = "phillychi3@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py (>=2.6.3,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.ruff]
lint.select = ["E", "F"]
lint.ignore = ["E501", "F401", "E711", "E712"]
lint.fixable = ["A", "B", "C", "D", "E", "F"]
lint.unfixable = []
target-version = "py312"
exclude = [".git", ".ruff_cache", ".venv", "venv"]

line-length = 120
